Measure the performance of iframe creation in/out of scope.
Usage: open nofetch-register.html to register a service worker, restart
your browser and run nofetch.html.

nofetch.html measures just creation of iframe, but the in-scope request
contains start-up, checking the fetch event handler and then falling
back to the network while the out-of-scope request directly goes to the
network.